v2-pipeline-LLM/requirements.txt
anyio 
argon2-cffi 
argon2-cffi-bindings 
asttokens 
attrs 
Babel 
backcall 
beautifulsoup4 
beautifultable==1.1.0
bertopic==0.14.1
bleach 
bokeh 
Bottleneck 
brotlipy==0.7.0
certifi 
cffi 
charset-normalizer 
click==8.1.3
cloudpickle 
cmake==3.26.0
comm 
contourpy 
cryptography 
cycler 
Cython==0.29.33
cytoolz 
dacite==1.8.1
dask 
debugpy 
decorator 
defusedxml 
distributed 
entrypoints 
et-xmlfile==1.1.0
executing 
fastjsonschema 
filelock==3.10.0
flit_core 
fonttools==4.25.0
fsspec 
hdbscan==0.8.29
HeapDict 
htmlmin==0.1.12
huggingface-hub==0.13.2
idna 
ImageHash==4.3.1
importlib-metadata 
ipykernel 
ipython 
ipython-genutils 
ipywidgets 
jedi 
Jinja2 
joblib==1.2.0
json5 
jsonschema 
jupyter 
jupyter-console 
jupyter-server 
jupyter_client 
jupyter_core 
jupyterlab 
jupyterlab-pygments 
jupyterlab-widgets 
jupyterlab_server 
kiwisolver 
langdetect==1.0.9
lit==15.0.7
llvmlite==0.39.1
locket 
lxml 
lz4 
MarkupSafe 
matplotlib 
matplotlib-inline 
mistune 
mkl-fft==1.3.1
mkl-random 
mkl-service==2.4.0
mpmath==1.3.0
msgpack 
multimethod==1.9.1
munkres==1.1.4
nbclassic 
nbclient 
nbconvert 
nbformat 
nest-asyncio 
networkx==3.0
nltk==3.8.1
notebook 
notebook_shim 
numba==0.56.4
numexpr 
numpy 
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openpyxl==3.1.2
packaging 
pandas==1.5.2
pandocfilters 
parso 
partd 
patsy==0.5.3
pexpect 
phik==0.12.3
pickleshare 
Pillow==9.4.0
platformdirs 
plotly 
ply==3.11
prometheus-client 
prompt-toolkit 
psutil 
ptyprocess 
pure-eval 
pyarrow==10.0.1
pycparser 
pydantic==1.10.9
Pygments 
pynndescent==0.5.8
pyOpenSSL 
pyparsing 
PyQt5-sip==12.11.0
pyrsistent 
PySocks 
python-dateutil 
pytz 
PyWavelets==1.4.1
PyYAML 
pyzmq 
qtconsole 
QtPy 
regex==2022.10.31
requests 
scikit-learn==1.2.2
scipy==1.10.1
seaborn 
Send2Trash 
sentence-transformers==2.2.2
sentencepiece==0.1.97
sip 
six 
sniffio 
sortedcontainers 
soupsieve 
stack-data 
statsmodels==0.14.0
sympy==1.11.1
tangled-up-in-unicode==0.2.0
tblib 
tenacity 
terminado 
threadpoolctl==3.1.0
tinycss2 
tokenizers==0.13.2
toml 
tomli 
toolz 
torch==2.0.0
torchvision==0.15.1
tornado 
tqdm 
traitlets 
transformers==4.27.1
triton==2.0.0
typeguard==2.13.3
typing_extensions 
umap-learn==0.5.3
urllib3 
visions==0.7.5
wcwidth 
webencodings==0.5.1
websocket-client 
widgetsnbextension 
wordcloud==1.9.2
ydata-profiling==4.2.0
zict==2.1.0
zipp 
